feat(ravioli): Support non observable representation
A component can now be reduce as a simple primitive or any other non
observable representation.
This can be usefull for server side scenario when a store won't be observed.
To enable this, simple set the 3rd argument as `false` in setTransformation() to tell the component thas representation won't be observed.
As a result you will need to trigger yourself the next instruction of the step (rerender, send the result, etc...)